fix warnings from unsigned promotion fix newer std::format stuff

+++ lib/types/allocator.cpp
@@ -2,7 +2,7 @@
 #include "lib/types/allocator.h"
 #include "lib/gi.h"
 
-uint32_t internal::non_game_count, internal::game_count;
+size_t internal::non_game_count, internal::game_count;
 
 void *internal::alloc(size_t len, mem_tag tag)
 {
